requirements.txt
questionary==2.1.1
rich==14.2.0
python-dotenv==1.2.1
openai==2.7.1
requests==2.32.5